perf(site): cache client JS/CSS, self-host fonts, fix jank
Speed up loads and eliminate stutter across the site:

- Navbar dock drag: snapshot tile geometry on pointerdown instead of
  reading getBoundingClientRect 2xN per pointermove (layout-thrash jank).
- Search: coalesce re-renders to one per frame (rAF) and skip rebuilds
  when the query is unchanged.
- Lazy-load below-fold images (game tiles, leaderboard/birthday avatars).
- Extract theme/navbar/search inline JS into a single cached, hash-busted
  /static/app.js (loaded deferred); the search index now ships as a
  non-executed JSON data island. HTML is no-store, so this stops re-parsing
  ~18KB of JS on every navigation. Adds generalized asset-version helper.
- Move navbar + search <style> blocks into the cached styles.css.
- Self-host Outfit + JetBrains Mono (latin-subset variable woff2) under
  /static/fonts and add @font-face with font-display: swap. The strict CSP
  (font-src 'self') was silently blocking the Google Fonts <link>, leaving
  the site on system fallback; this fixes that and drops a render-blocking
  third-party request + two dead preconnects.
- eslint: ignore site_src/Assets (static browser asset, not server source).
