+++ site_src/components/navbar.ts
@@ -52,392 +52,7 @@ export interface NavUser {
   csrf: string;
 }
 
-const navbarStyles = raw(`
-<style>
-  /* Default (desktop / non-touch): top links visible, mobile dock hidden. */
-  #nav-links  { display: flex; }
-  #nav-mobile { display: none; }
-
-  .nav-link.active { border-bottom-color: var(--accent); }
-  .nav-links.js-ready .nav-link.active { border-bottom-color: transparent; }
-
-  /* Caisena gradient nav surfaces — theme-aware via CSS vars */
-  .nav-surface {
-    background: color-mix(in oklab, var(--ink-800) 75%, transparent);
-    backdrop-filter: blur(16px);
-    -webkit-backdrop-filter: blur(16px);
-    border-bottom: 1px solid transparent;
-    border-image: linear-gradient(90deg, transparent 0%, var(--accent) 30%, var(--accent-pale) 70%, transparent 100%) 1;
-    box-shadow: 0 4px 30px rgba(0, 0, 0, 0.3), inset 0 1px 0 rgba(255, 255, 255, 0.05);
-  }
-  .nav-underline-grad {
-    background: linear-gradient(90deg, var(--accent) 0%, var(--accent-pale) 100%);
-    box-shadow: 0 0 14px var(--glow-bright), 0 0 4px var(--accent);
-  }
-
-  /* Desktop window header is hidden by default (mobile/narrow) */
-  .nav-window-header {
-    display: none;
-  }
-  @media (min-width: 1025px) {
-    .nav-window-header {
-      display: flex;
-    }
-  }
-
-  /* Desktop window styles override */
-  @media (min-width: 1025px) {
-    .nav-surface {
-      max-width: 1100px;
-      width: calc(100% - 2rem);
-      margin: 1.5rem auto 0 auto;
-      border-radius: 0.75rem;
-      border: 1px solid color-mix(in oklab, var(--accent) 20%, transparent);
-      border-image: none;
-      background: color-mix(in oklab, var(--ink-800) 65%, transparent);
-      backdrop-filter: blur(12px);
-      -webkit-backdrop-filter: blur(12px);
-      box-shadow: 
-        0 10px 30px rgba(0, 0, 0, 0.4),
-        inset 0 1px 0 rgba(255, 255, 255, 0.05),
-        0 0 15px rgba(34, 211, 255, 0.05);
-      position: relative;
-      overflow: hidden;
-    }
-  }
-
-  /* Desktop nav link spacing */
-  #nav-links { gap: 1.75rem; }
-
-  /* Spacer balances the logo so links stay visually centered when no auth chip is present. */
-  .nav-spacer { width: 3rem; height: 3rem; flex: none; }
-
-  /* Desktop auth chip (right side of navbar, logged-in only) */
-  #nav-auth-desktop { display: flex; align-items: center; }
-  .nav-auth { display: flex; align-items: center; gap: 0.6rem; color: var(--fog-200); font-size: 0.9rem; font-family: 'JetBrains Mono', monospace !important; }
-  /* Avatar + username are wrapped in <a href="/me"> — the link itself stays unstyled. */
-  .nav-profile-link { display: inline-flex; align-items: center; gap: 0.6rem; text-decoration: none; color: inherit; }
-  .nav-avatar { width: 32px; height: 32px; border-radius: 50%; border: 1px solid var(--accent); transition: box-shadow 0.2s; }
-  .nav-username { font-weight: 500; color: var(--fog-100); max-width: 10rem; overflow: hidden; text-overflow: ellipsis; white-space: nowrap; text-decoration: none; }
-  .nav-profile-link:hover .nav-username { text-decoration: underline; text-underline-offset: 3px; text-decoration-color: var(--accent); }
-  .nav-profile-link:hover .nav-avatar { box-shadow: 0 0 6px var(--glow-faint, rgba(255,255,255,0.2)); }
-  .nav-auth-link { color: color-mix(in oklab, var(--danger) 65%, #000); text-decoration: none; font-size: 0.85rem; padding: 0.3rem 0.6rem; border-radius: 0.4rem; border: 1px solid transparent; transition: color 0.2s, text-shadow 0.2s; }
-  .nav-auth-link:hover { color: var(--danger); text-shadow: 0 0 8px var(--danger-glow); }
-  .nav-login-link { color: color-mix(in oklab, var(--accent) 65%, #000); text-decoration: none; font-size: 0.85rem; padding: 0.3rem 0.6rem; border-radius: 0.4rem; border: 1px solid transparent; transition: color 0.2s, text-shadow 0.2s; }
-  .nav-login-link:hover { color: var(--accent); text-shadow: 0 0 8px var(--glow-bright); }
-
-  /* ── Touch / narrow-viewport dock ───────────────────────────────────────
-     Floating glass tab bar pinned to the bottom. Triggered by either a
-     narrow viewport (covers tablets, foldables, phones in DevTools where
-     hover/pointer emulation is unreliable) OR a true touch device (covers
-     touchscreens that happen to be wider, e.g. Nest Hub Max, iPad Pro
-     12.9 in landscape). */
-  @media (max-width: 1024px), (hover: none) and (pointer: coarse) {
-    #nav-links { display: none; }
-
-    /* Reserve scroll space below the page content so the floating dock
-       never permanently covers the footer — users can scroll past it. */
-    body {
-      padding-bottom: calc(4.5rem + env(safe-area-inset-bottom));
-    }
-
-    #nav-mobile {
-      display: flex !important;
-      position: fixed;
-      bottom: calc(0.9rem + env(safe-area-inset-bottom));
-      left: 50%;
-      transform: translateX(-50%);
-      width: min(92vw, 26rem);
-      padding: 0.45rem 0.5rem;
-      border-radius: 9999px;
-      background: color-mix(in oklab, var(--ink-900) 72%, transparent) !important;
-      -webkit-backdrop-filter: blur(25px) saturate(190%);
-      backdrop-filter: blur(25px) saturate(190%);
-      border: 1px solid color-mix(in oklab, var(--accent) 35%, transparent) !important;
-      border-image: none !important;
-      box-shadow:
-        0 12px 32px rgba(0, 0, 0, 0.6),
-        inset 0 1px 0 rgba(255, 255, 255, 0.08),
-        0 0 15px rgba(34, 211, 255, 0.15);
-      flex-direction: row;
-      justify-content: space-around;
-      align-items: stretch;
-      gap: 0;
-      z-index: 50;
-      isolation: isolate;
-    }
-
-    /* Sliding pill — accent-tinted, animates between tiles, draggable */
-    #nav-mobile .dock-pill {
-      position: absolute;
-      top: 0.4rem;
-      bottom: 0.4rem;
-      left: 0;
-      width: 0;
-      border-radius: 9999px;
-      background: linear-gradient(90deg, color-mix(in oklab, var(--accent) 25%, transparent), color-mix(in oklab, var(--accent-pale) 25%, transparent));
-      border: 1px solid color-mix(in oklab, var(--accent) 45%, transparent);
-      box-shadow:
-        inset 0 1px 0 rgba(255, 255, 255, 0.15),
-        0 0 12px rgba(34, 211, 255, 0.25);
-      pointer-events: auto;
-      touch-action: none;
-      cursor: grab;
-      z-index: 0;
-      will-change: transform, width;
-    }
-    #nav-mobile .dock-pill.dragging {
-      cursor: grabbing;
-      box-shadow:
-        inset 0 1px 0 rgba(255, 255, 255, 0.5),
-        inset 0 -1px 0 rgba(0, 0, 0, 0.1),
-        0 4px 16px color-mix(in oklab, var(--accent) 40%, transparent);
-    }
-
-    /* Tile = icon + label, stacked. Sits above the pill in z-stack. */
-    #nav-mobile .nav-link {
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      justify-content: center;
-      gap: 3px;
-      padding: 0.45rem 0.25rem;
-      border: 0 !important;
-      border-radius: 9999px;
-      background: transparent !important;
-      color: var(--fog-300);
-      position: relative;
-      z-index: 1;
-      transition: color 0.32s cubic-bezier(0.4, 0, 0.2, 1);
-    }
-    #nav-mobile .nav-link svg {
-      width: 22px;
-      height: 22px;
-      display: block;
-    }
-    #nav-mobile .nav-link .label {
-      font-size: 0.62rem;
-      line-height: 1;
-      letter-spacing: 0.02em;
-      font-weight: 500;
-    }
-    /* Active state: text color brightens; the pill below provides the bg */
-    #nav-mobile .nav-link.active { color: var(--fog-100); }
-  }
-</style>
-`);
-
-const navbarScript = (nonce: string) => raw(`
-<script nonce="${nonce}">
-(() => {
-  // ── Desktop sliding underline ──────────────────────────────────────────────
-  const container = document.getElementById('nav-links');
-  if (container) {
-    const underline = container.querySelector('.nav-underline');
-    const links = Array.from(container.querySelectorAll('.nav-link'));
-    if (underline && links.length) {
-      container.classList.add('js-ready');
-
-      const place = (el, animate) => {
-        const cRect = container.getBoundingClientRect();
-        const r = el.getBoundingClientRect();
-        underline.style.transition = animate
-          ? 'transform 0.28s cubic-bezier(0.4, 0, 0.2, 1), width 0.28s cubic-bezier(0.4, 0, 0.2, 1)'
-          : 'none';
-        underline.style.width = r.width + 'px';
-        underline.style.transform = 'translateX(' + (r.left - cRect.left) + 'px)';
-        underline.style.opacity = '1';
-      };
-
-      const activeLink = links.find((l) => l.classList.contains('active'));
-      if (activeLink) requestAnimationFrame(() => place(activeLink, false));
-
-      links.forEach((l) => {
-        l.addEventListener('click', (e) => {
-          if (l.classList.contains('active')) return;
-          if (e.metaKey || e.ctrlKey || e.shiftKey || e.altKey || e.button === 1) return;
-          e.preventDefault();
-          place(l, true);
-          setTimeout(() => { window.location.href = l.href; }, 240);
-        });
-      });
-
-      window.addEventListener('resize', () => {
-        const current = links.find((x) => x.classList.contains('active'));
-        if (current) place(current, false);
-      });
-
-      const theme = new URLSearchParams(location.search).get('theme');
-      if (theme) {
-        links.forEach((l) => {
-          const u = new URL(l.href);
-          u.searchParams.set('theme', theme);
-          l.href = u.toString();
-        });
-      }
-    }
-  }
-
-  // ── Mobile dock: sliding pill + drag-to-focus ─────────────────────────────
-  const dock = document.getElementById('nav-mobile');
-  const pill = dock && dock.querySelector('.dock-pill');
-  // Mirror the CSS that decides when the dock is visible. Evaluate the two
-  // queries separately — older WebKit rejects comma-joined media-query lists
-  // passed to matchMedia.
-  const dockVisible = !!window.matchMedia && (
-    window.matchMedia('(max-width: 1024px)').matches
-    || window.matchMedia('(hover: none) and (pointer: coarse)').matches
-  );
-  if (dock && pill && dockVisible) {
-    const tiles = Array.from(dock.querySelectorAll('.nav-link'));
-    const NAV_DELAY = 320;
-    const SPRING = '0.4s cubic-bezier(0.34, 1.56, 0.64, 1)';
-
-    let activeIdx = tiles.findIndex((t) => t.classList.contains('active'));
-    if (activeIdx < 0) activeIdx = 0;
-    let pillX = 0;
-    let pillW = 0;
-    let pendingNav = null;
-
-    const tileGeom = (tile) => {
-      const dr = dock.getBoundingClientRect();
-      const tr = tile.getBoundingClientRect();
-      return { x: tr.left - dr.left, w: tr.width };
-    };
-
-    const placePill = (idx, animate) => {
-      const g = tileGeom(tiles[idx]);
-      pillX = g.x; pillW = g.w;
-      pill.style.transition = animate ? 'transform ' + SPRING + ', width ' + SPRING : 'none';
-      pill.style.width = g.w + 'px';
-      pill.style.transform = 'translateX(' + g.x + 'px)';
-    };
-
-    const setActive = (idx) => {
-      if (idx === activeIdx) return;
-      tiles[activeIdx].classList.remove('active');
-      tiles[activeIdx].removeAttribute('aria-current');
-      tiles[idx].classList.add('active');
-      tiles[idx].setAttribute('aria-current', 'page');
-      activeIdx = idx;
-    };
-
-    const navigate = (idx) => {
-      if (pendingNav) clearTimeout(pendingNav);
-      pendingNav = setTimeout(() => { window.location.href = tiles[idx].href; }, NAV_DELAY);
-    };
-
-    requestAnimationFrame(() => placePill(activeIdx, false));
-
-    // Tap a tile → slide pill + navigate (skip if it was a drag-induced click)
-    tiles.forEach((tile, i) => {
-      tile.addEventListener('click', (e) => {
-        if (suppressClick) { suppressClick = false; e.preventDefault(); e.stopPropagation(); return; }
-        if (i === activeIdx) return;
-        if (e.metaKey || e.ctrlKey || e.shiftKey || e.altKey || e.button === 1) return;
-        e.preventDefault();
-        setActive(i);
-        placePill(i, true);
-        navigate(i);
-      });
-    });
-
-    // ── Drag the pill to slide focus ─────────────────────────────
-    let dragging = false;
-    let dragStartX = 0;
-    let dragOriginX = 0;
-    let dragMoved = false;
-    let suppressClick = false;
-
-    dock.addEventListener('pointerdown', (e) => {
-      // Only initiate drag when the pointer lands on the pill itself.
-      const dr = dock.getBoundingClientRect();
-      const localX = e.clientX - dr.left;
-      if (localX < pillX || localX > pillX + pillW) return;
-      dragging = true;
-      dragMoved = false;
-      dragStartX = e.clientX;
-      dragOriginX = pillX;
-      pill.classList.add('dragging');
-      pill.style.transition = 'none';
-      try { dock.setPointerCapture(e.pointerId); } catch (_) {}
-    });
-
-    dock.addEventListener('pointermove', (e) => {
-      if (!dragging) return;
-      const dx = e.clientX - dragStartX;
-      if (Math.abs(dx) > 4) dragMoved = true;
-      const dr = dock.getBoundingClientRect();
-      // Pad inside the dock so the pill never overlaps the curved edge
-      const pad = 4;
-      const minX = pad;
-      const maxX = dr.width - pillW - pad;
-      let nx = dragOriginX + dx;
-      if (nx < minX) nx = minX;
-      if (nx > maxX) nx = maxX;
-      pillX = nx;
-      pill.style.transform = 'translateX(' + nx + 'px)';
-
-      // Live preview: brighten whichever tile the pill center is closest to
-      const center = nx + pillW / 2;
-      let nearest = 0;
-      let minDist = Infinity;
-      tiles.forEach((t, i) => {
-        const g = tileGeom(t);
-        const c = g.x + g.w / 2;
-        const d = Math.abs(c - center);
-        if (d < minDist) { minDist = d; nearest = i; }
-      });
-      tiles.forEach((t, i) => t.classList.toggle('active', i === nearest));
-    });
-
-    const endDrag = (e) => {
-      if (!dragging) return;
-      dragging = false;
-      pill.classList.remove('dragging');
-      try { dock.releasePointerCapture(e.pointerId); } catch (_) {}
-      if (!dragMoved) {
-        // No real drag — restore pill under the previously-active tile.
-        tiles.forEach((t, i) => t.classList.toggle('active', i === activeIdx));
-        if (activeIdx >= 0) tiles[activeIdx].setAttribute('aria-current', 'page');
-        placePill(activeIdx, true);
-        return;
-      }
-      // Snap to nearest tile, navigate if changed.
-      suppressClick = true;
-      setTimeout(() => { suppressClick = false; }, 400);
-      const center = pillX + pillW / 2;
-      let nearest = 0;
-      let minDist = Infinity;
-      tiles.forEach((t, i) => {
-        const g = tileGeom(t);
-        const c = g.x + g.w / 2;
-        const d = Math.abs(c - center);
-        if (d < minDist) { minDist = d; nearest = i; }
-      });
-      const changed = nearest !== activeIdx;
-      // Update aria-current on the new active tile
-      tiles.forEach((t, i) => {
-        t.classList.toggle('active', i === nearest);
-        if (i === nearest) t.setAttribute('aria-current', 'page');
-        else t.removeAttribute('aria-current');
-      });
-      activeIdx = nearest;
-      placePill(nearest, true);
-      if (changed) navigate(nearest);
-    };
-
-    dock.addEventListener('pointerup', endDrag);
-    dock.addEventListener('pointercancel', endDrag);
-
-    window.addEventListener('resize', () => placePill(activeIdx, false));
-  }
-})();
-</script>
-`);
-
-export function Navbar(active: NavActive | undefined, nonce: string, lv999?: boolean, user?: NavUser | null) {
+export function Navbar(active: NavActive | undefined, lv999?: boolean, user?: NavUser | null) {
   const base = 'nav-link text-[0.95rem] px-[0.1rem] py-1 border-b-2 border-transparent transition-colors no-underline font-mono';
   const link = (href: string, label: string, key: string) => {
     const isActive = active === key;
@@ -459,7 +74,6 @@ export function Navbar(active: NavActive | undefined, nonce: string, lv999?: boo
   const sticker = pool[Math.floor(Math.random() * pool.length)];
 
   return html`
-    ${navbarStyles}
     <nav id="site-nav" class="nav-surface font-mono">
       <!-- Desktop window header: only visible on desktop -->
       <div class="nav-window-header flex items-center justify-between px-[1.5rem] py-2 border-b border-[rgba(34,211,255,0.12)] select-none">
@@ -519,7 +133,5 @@ export function Navbar(active: NavActive | undefined, nonce: string, lv999?: boo
       ${dockLink('/birthdays', 'Birthdays', 'birthdays')}
       ${dockLink('/games', 'Games', 'games')}
     </div>
-
-    ${navbarScript(nonce)}
   `;
 }
